Changed Camera to inherit from type IComponent, Updated Camera to use a Aspect Ration when calculating view
- Viewport in editor now updates the main camera views aspect ratio
- Application now stores a main camera entity as camera now must have an entity in order to get it's position
- Can now scroll in in editor

+++ ChaosEngine/src/Chaos/Renderer/Renderer.h
@@ -10,8 +10,8 @@ namespace Chaos
 {
 	struct DebugInfo
 	{
-		int TotalQuadsDrawn = 0;
-		int NumOfDrawCalls = 0;
+		size_t TotalQuadsDrawn = 0;
+		uint32_t NumOfDrawCalls = 0;
 	};
 
 	class Renderer
